Ref: Created user_balances and moved files inside
