+++ app/src/components/market/sections/common_sections/user_balances/asset_balance/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { Token } from '../../../../util/types'
+import { Token } from '../../../../../../util/types'
 
 const Wrapper = styled.div`
   align-items: center;
